Main/requirements.txt
Flask
Flask-SQLAlchemy
Werkzeug
SQLAlchemy
APScheduler
python-dotenv
Flask-Migrate
gunicorn
pymysql
flask-wtf